+++ ansible_mitogen/connection.py
@@ -890,6 +890,29 @@ class Connection(ansible.plugins.connection.ConnectionBase):
             self.binding.close()
             self.binding = None
 
+    def _mitogen_var_options(self, templar):
+        # Workaround for https://github.com/ansible/ansible/issues/84238
+        var_names = C.config.get_plugin_vars('connection', self._load_name)
+        variables = templar.available_variables
+        var_options = {
+            var_name: templar.template(variables[var_name])
+            for var_name in var_names
+            if var_name in variables
+        }
+
+        if self.allow_extras:
+            extras_var_prefix = 'ansible_%s_' % self.extras_prefix
+            var_options['_extras'] = {
+                var_name: templar.template(variables[var_name])
+                for var_name in variables
+                if var_name not in var_options
+                and var_name.startswith(extras_var_prefix)
+            }
+        else:
+            var_options['_extras'] = {}
+
+        return var_options
+
     reset_compat_msg = (
         'Mitogen only supports "reset_connection" on Ansible 2.5.6 or later'
     )
@@ -921,6 +944,19 @@ class Connection(ansible.plugins.connection.ConnectionBase):
             templar=0,
             shared_loader_obj=0
         )
+
+        # Workaround for https://github.com/ansible/ansible/issues/84238
+        try:
+            task, templar = self._play_context.vars.pop(
+                '_mitogen.smuggled.reset_connection',
+            )
+        except KeyError:
+            pass
+        else:
+            self.set_options(
+                task_keys=task.dump_attrs(),
+                var_options=self._mitogen_var_options(templar),
+            )
 
         # Clear out state in case we were ever connected.
         self.close()